Images fix

+++ src/main/java/com/heroku/spacey/services/impl/AwsImageServiceImpl.java
@@ -82,8 +82,14 @@ public class AwsImageServiceImpl implements AwsImageService {
 
     @Override
     public void delete(String url) {
+        if (url == null || url.isEmpty()) {
+            log.warn("null or empty url for photo");
+            return;
+        }
+
         if (!s3client.doesObjectExist(bucketName, getPhotoNameByUrl(url))) {
-            throw new NotFoundException("image not found in AWSs3 cloud");
+            log.warn("image not found in AWSs3 cloud");
+            return;
         }
         s3client.deleteObject(bucketName, getPhotoNameByUrl(url));
         log.info("image deleted from AWS");
